2026_01_08_VectorDB-Search/requirements.txt
fastapi>=0.110
uvicorn>=0.27
pymongo>=4.6
pydantic-settings>=2.0
python-dotenv>=1.0
ollama
pypdf

pytest
httpx
